fix(etapi): mask protected attachment titles in the attachment mapper
`mapAttachmentToPojo` read `attachment.title` directly, so a protected
attachment's raw base64 AES ciphertext title was returned whenever the
protected session was locked — the permanent state for ETAPI.

`BAttachment.decrypt()` leaves the ciphertext in `title` when there is no
session, so the raw read publishes it. `getTitleOrProtected()` guards
that, and behaves identically to the note mapper: `isContentAvailable()`
returns the real title for unprotected attachments and for protected ones
while unlocked. Newly created attachments are unaffected — they have no
attachmentId yet, which `isContentAvailable()` already treats as available.

This covers both the listing and the by-id read. They are one defect
reached through two addresses: `GET /etapi/notes/:noteId/attachments`
and `GET /etapi/attachments/:attachmentId` map through this same
function, and both returned the identical ciphertext.

`mime` and `contentLength` stay in the payload, matching the note mapper,
which keeps unencrypted metadata such as attribute values.

The spec covering the note mapper is renamed, since it now also covers
attachments.

Reported as GHSA-2g36-rhpj-xxc9 and GHSA-cj5g-3c55-8g2c.
