+++ apps/server/src/assets/etapi.openapi.yaml
@@ -1184,6 +1184,9 @@ components:
                     type: string
                 title:
                     type: string
+                    description: |
+                        Title of the attachment. Returned as "[protected]" for protected attachments,
+                        since ETAPI cannot open a protected session.
                 position:
                     type: integer
                     format: int32
